+++ src/bin/splitfs.rs
@@ -1,5 +1,5 @@
 use scfs::cli::Cli;
 
 fn main() {
-    Cli::SCFS.run();
+    Cli::SplitFS.run();
 }