+++ src/catfs.rs
@@ -13,9 +13,10 @@ use libc::ENOENT;
 use rusqlite::{params, Connection, Error, NO_PARAMS};
 
 use crate::{
-    convert_metadata_to_attr, Config, FileHandle, FileInfo, FileInfoRow, CONFIG_FILE_NAME,
-    INO_OUTSIDE, INO_ROOT, STMT_CREATE, STMT_CREATE_INDEX_PARENT_INO_FILE_NAME, STMT_INSERT,
-    STMT_QUERY_BY_INO, STMT_QUERY_BY_PARENT_INO, STMT_QUERY_BY_PARENT_INO_AND_FILENAME, TTL,
+    convert_filetype, convert_metadata_to_attr, Config, FileHandle, FileInfo, FileInfoRow,
+    CONFIG_FILE_NAME, INO_OUTSIDE, INO_ROOT, STMT_CREATE, STMT_CREATE_INDEX_PARENT_INO_FILE_NAME,
+    STMT_INSERT, STMT_QUERY_BY_INO, STMT_QUERY_BY_PARENT_INO,
+    STMT_QUERY_BY_PARENT_INO_AND_FILENAME, TTL,
 };
 
 pub struct CatFS {
@@ -122,7 +123,10 @@ impl CatFS {
             let attrs = parts
                 .iter()
                 .map(|info| {
-                    convert_metadata_to_attr(fs::metadata(&info.path).unwrap(), Some(info.ino))
+                    convert_metadata_to_attr(
+                        fs::symlink_metadata(&info.path).unwrap(),
+                        Some(info.ino),
+                    )
                 })
                 .collect::<Vec<_>>();
             let mut attr = attrs.get(0).unwrap().clone();
@@ -132,7 +136,7 @@ impl CatFS {
             attr
         } else {
             let attr = convert_metadata_to_attr(
-                fs::metadata(&file_info.path).unwrap(),
+                fs::symlink_metadata(&file_info.path).unwrap(),
                 Some(file_info.ino),
             );
             attr
@@ -141,6 +145,14 @@ impl CatFS {
 
     fn populate<P: AsRef<Path>>(file_db: &Connection, path: P, parent_ino: u64) {
         let path = path.as_ref();
+
+        let meta = path.symlink_metadata().unwrap();
+
+        if let None = convert_filetype(meta.file_type()) {
+            return;
+        }
+
+        let attr = convert_metadata_to_attr(meta, None);
 
         if path.file_name().unwrap() == CONFIG_FILE_NAME {
             return;
@@ -157,7 +169,7 @@ impl CatFS {
             parent_ino,
             path: OsString::from(path),
             file_name: path.file_name().unwrap().into(),
-            part: if path.is_file() {
+            part: if let FileType::RegularFile = attr.kind {
                 path.file_name().unwrap().to_str().unwrap()[5..]
                     .parse::<u64>()
                     .unwrap()
@@ -166,6 +178,7 @@ impl CatFS {
                 0
             },
             vdir: false,
+            symlink: attr.kind == FileType::Symlink,
         });
 
         file_db
@@ -177,11 +190,12 @@ impl CatFS {
                 file_info.path,
                 file_info.file_name,
                 file_info.part,
-                file_info.vdir
+                file_info.vdir,
+                file_info.symlink,
             ])
             .unwrap();
 
-        if path.is_dir() {
+        if let FileType::Directory = attr.kind {
             for entry in fs::read_dir(path).unwrap() {
                 let entry = entry.unwrap();
                 CatFS::populate(&file_db, entry.path(), ino);
@@ -210,6 +224,13 @@ impl Filesystem for CatFS {
         } else {
             reply.error(ENOENT)
         }
+    }
+
+    fn readlink(&mut self, _req: &Request, ino: u64, reply: ReplyData) {
+        let path = self.get_file_info_from_ino(ino).unwrap().path;
+        let target = fs::read_link(path).unwrap();
+        let target = target.to_str().unwrap().as_bytes();
+        reply.data(target);
     }
 
     fn open(&mut self, _req: &Request, ino: u64, _flags: u32, reply: ReplyOpen) {
@@ -380,7 +401,9 @@ impl Filesystem for CatFS {
                         }
                         + off as i64
                         + 1,
-                    if item.vdir {
+                    if item.symlink {
+                        FileType::Symlink
+                    } else if item.vdir {
                         FileType::RegularFile
                     } else {
                         FileType::Directory
@@ -404,6 +427,7 @@ mod tests {
     use std::io::Write;
     use std::iter;
     use std::ops::Deref;
+    use std::os::unix::fs::symlink;
 
     use fuse::BackgroundSession;
     use rand::{thread_rng, Rng, RngCore};
@@ -422,9 +446,10 @@ mod tests {
         pub(crate) mountpoint: TempDir,
     }
 
-    fn mount_and_create_files<'a>(
+    fn mount_and_create_files_with_symlinks<'a>(
         files: &Vec<(String, Vec<u8>)>,
-    ) -> Result<(TempSession<'a>), std::io::Error> {
+        symlinks: Vec<(String, String)>,
+    ) -> Result<TempSession<'a>, std::io::Error> {
         let mirror = tempdir()?;
         let mountpoint = tempdir()?;
 
@@ -435,15 +460,25 @@ mod tests {
             file.write_all(&data)?;
         }
 
+        for (link_name, target) in symlinks {
+            symlink(&target, mirror.path().join(&link_name))?;
+        }
+
         let fs = CatFS::new(mirror.path().as_os_str());
 
-        let session = mount(fs, &mountpoint);
+        let session = mount(fs, &mountpoint, Vec::new());
 
         Ok(TempSession {
             _mirror: mirror,
             mountpoint,
             _session: session,
         })
+    }
+
+    fn mount_and_create_files<'a>(
+        files: &Vec<(String, Vec<u8>)>,
+    ) -> Result<TempSession<'a>, std::io::Error> {
+        mount_and_create_files_with_symlinks(files, Vec::new())
     }
 
     fn create_random_file_tuples(
@@ -498,10 +533,6 @@ mod tests {
             .map(|entry| entry.unwrap().path())
             .collect::<Vec<_>>();
 
-        for (file_name, content) in &files_expected {
-            println!("{}: {}", file_name, content.len());
-        }
-
         let mut files_expected_map = HashMap::new();
         for (file_name, content_chunk) in files_expected {
             if file_name == CONFIG_FILE_NAME {
@@ -535,7 +566,7 @@ mod tests {
     fn test_empty_mirror() {
         // Since a valid SplitFS needs a config file, panic if there is no such file
 
-        let session = mount_and_create_files(&vec![]).unwrap();
+        let session = mount_and_create_files(&Vec::new()).unwrap();
 
         let entries = fs::read_dir(&session.mountpoint)
             .unwrap()
@@ -630,5 +661,150 @@ mod tests {
         let session = mount_and_create_files(&files_expected)?;
 
         check_files(session.mountpoint.path(), files_expected)
+    }
+
+    #[test]
+    fn test_symlink_relative_file() -> Result<(), std::io::Error> {
+        // A symlink should just be presented as such. A
+        // relative symlink pointing to a file inside the chunked directory should translate into
+        // the same concatenated file as the real file.
+
+        let config = Config::default().blocksize(1);
+        let blocksize = config.blocksize as usize;
+        let num_files = 1;
+        let max_num_fragments = 5;
+
+        let files = with_config_file(
+            create_random_file_tuples(blocksize, num_files, max_num_fragments),
+            config,
+        );
+
+        let mut symlink_map = HashMap::new();
+        symlink_map.insert(
+            "link_rel".into(),
+            files.first().unwrap().0.split("/").next().unwrap().into(),
+        );
+
+        let session = mount_and_create_files_with_symlinks(
+            &files,
+            symlink_map.clone().into_iter().collect(),
+        )?;
+
+        let entries = fs::read_dir(&session.mountpoint)?
+            .map(|entry| entry.unwrap())
+            .collect::<Vec<_>>();
+
+        assert_eq!(entries.len(), num_files + symlink_map.len());
+
+        let symlinks_found = entries
+            .iter()
+            .filter(|item| item.file_type().unwrap().is_symlink())
+            .collect::<Vec<_>>();
+
+        assert_eq!(symlinks_found.len(), symlink_map.len());
+
+        let contents = entries
+            .iter()
+            .map(|file| fs::read(file.path()).unwrap())
+            .collect::<Vec<_>>();
+
+        assert!(contents
+            .iter()
+            .all(|content| content.eq(contents.first().unwrap())));
+
+        for symlink in symlinks_found {
+            let symlink = symlink.path();
+            let link_name = symlink.file_name().unwrap().to_str().unwrap();
+            let target = symlink_map.remove(link_name);
+            assert_ne!(target, None);
+            let target = target.unwrap();
+            assert_eq!(fs::read_link(symlink)?.to_str().unwrap(), target);
+        }
+
+        assert!(symlink_map.is_empty());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_symlink_absolute_dir() -> Result<(), std::io::Error> {
+        // A symlink should just be presented as such.
+
+        let mut symlink_map = HashMap::new();
+        symlink_map.insert("link_abs".into(), "/".into());
+
+        let session = mount_and_create_files_with_symlinks(
+            &with_config_file(Vec::new(), Config::default()),
+            symlink_map.clone().into_iter().collect(),
+        )?;
+
+        let entries = fs::read_dir(&session.mountpoint)?
+            .map(|entry| entry.unwrap())
+            .collect::<Vec<_>>();
+
+        assert_eq!(entries.len(), symlink_map.len());
+
+        let symlinks_found = entries
+            .iter()
+            .filter(|item| item.file_type().unwrap().is_symlink())
+            .collect::<Vec<_>>();
+
+        assert_eq!(symlinks_found.len(), symlink_map.len());
+
+        assert!(symlinks_found.first().unwrap().path().is_dir());
+
+        for symlink in symlinks_found {
+            let symlink = symlink.path();
+            let link_name = symlink.file_name().unwrap().to_str().unwrap();
+            let target = symlink_map.remove(link_name);
+            assert_ne!(target, None);
+            let target = target.unwrap();
+            assert_eq!(fs::read_link(symlink)?.to_str().unwrap(), target);
+        }
+
+        assert!(symlink_map.is_empty());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_symlink_broken() -> Result<(), std::io::Error> {
+        // A symlink should just be presented as such. If
+        // the target does not exist, just show a broken symlink, do not panic out.
+
+        let mut symlink_map = HashMap::new();
+        symlink_map.insert("link_rel".into(), "a/b/c".into());
+        symlink_map.insert("link_abs".into(), "/home/nobody/nothing".into());
+
+        let session = mount_and_create_files_with_symlinks(
+            &with_config_file(Vec::new(), Config::default()),
+            symlink_map.clone().into_iter().collect(),
+        )?;
+
+        let entries = fs::read_dir(&session.mountpoint)?
+            .map(|entry| entry.unwrap())
+            .collect::<Vec<_>>();
+
+        assert_eq!(entries.len(), symlink_map.len());
+
+        let symlinks_found = entries
+            .iter()
+            .filter(|item| item.file_type().unwrap().is_symlink())
+            .collect::<Vec<_>>();
+
+        assert_eq!(symlinks_found.len(), symlink_map.len());
+
+        for symlink in symlinks_found {
+            let symlink = symlink.path();
+            let link_name = symlink.file_name().unwrap().to_str().unwrap();
+            let target = symlink_map.remove(link_name);
+            assert_ne!(target, None);
+            let target = target.unwrap();
+            assert_eq!(fs::read_link(symlink)?.to_str().unwrap(), target);
+        }
+
+        assert!(symlink_map.is_empty());
+
+        Ok(())
     }
 }